Add comprehensive README for ADNL Tunnel project
A detailed README file has been added for the ADNL Tunnel project. This includes an overview of the service, a list of features, system architecture diagrams and explanations, prerequisites for use, installation instructions, configuration details and more. The aim is to provide users with all necessary information about the service and its usage in one place.